Fix type system (I think)
It is much more intelligent now, using all available information to infer types, including using the type of the data passed to functions to determine the function type. The type environment is now a single stack used for the whole typing process, which means that updates to the types of variables are reflected across the whole scope of the variable, and it is also more efficient than before.